brain_exploration: reduced size of behaviour tree
icp_node: added logic to get reference scans if multiple scans are bad in a row. Also added topic where map-odom tf is published, so that drive_2 can subscribe to it and use it when driving to not have to look it up each time.

drive_2: changed logic to lookup transform and added callback to set tf from map to odom using a subscriber to a topic from icp node.